+++ innobase/mtr/mtr0mtr.c
@@ -263,11 +263,11 @@ mtr_first_to_modify_page_after_backup(
 							backup_lsn) <= 0) {
 
 				printf("Page %lu newest %lu backup %lu\n",
-					block->offset,
-					ut_dulint_get_low(
+					(ulong) block->offset,
+					(ulong) ut_dulint_get_low(
 					buf_frame_get_newest_modification(
 							block->frame)),
-					ut_dulint_get_low(backup_lsn));
+					(ulong) ut_dulint_get_low(backup_lsn));
 					
 				ret = TRUE;
 			}
@@ -519,6 +519,6 @@ mtr_print(
 {
 	printf(
 	"Mini-transaction handle: memo size %lu bytes log size %lu bytes\n",
-		dyn_array_get_data_size(&(mtr->memo)),
-		dyn_array_get_data_size(&(mtr->log)));
+		(ulong) dyn_array_get_data_size(&(mtr->memo)),
+		(ulong) dyn_array_get_data_size(&(mtr->log)));
 }
